Capture fewer errors to reduce memory usage
On JRuby, with Nokogiri's default HTML parsing options ('recover'
instead of 'strict'), the `NokogiriNonStrictErrorHandler4NekoHtml` error
handler is used. We have found that this error handler can log a lot of
errors when encountering large docs with unusually pathological markup.

These error objects can consume a lot of memory, which might be OK if we
needed to refer to them, but in practice we don't. So it seems safe to
not capture them, and that way we benefit from a huge reduction in heap
usage in these problem-doc situations.